refactor: 로그인시 user Image가 null 이면 기본 이미지를 setUser하도록 수정 (#161)
* refactor: 로그인시 user Image가 null 이면 기본 이미지를 setUser하도록 수정

* chore: 기본 이미지 등록
