+++ src/constants/image.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_USER_IMAGE =
+  'https://user-images.githubusercontent.com/79133602/183457197-175f637c-c28b-4ddc-8cf1-195796fc3231.png'
