+++ src/recoil/currentUser.ts
@@ -1,16 +1,17 @@
 import { atom } from 'recoil'
 import { User } from '@interfaces'
 import { v1 } from 'uuid'
+import { DEFAULT_USER_IMAGE } from '@constants/image'
+
 export const currentUser = atom<User>({
   key: `currentUser/${v1()}`,
   default: {
     id: 1,
-    image:
-      'https://user-images.githubusercontent.com/79133602/181918487-b0e0d98c-3520-40f2-947b-7ba9f4422cd4.PNG',
-    email: 'example@naver.com',
-    nickName: '계란이 좋아',
-    snsAccount: 'example@naver.com',
-    createdAt: '2022-07-28T15:10:59.2186394',
+    image: DEFAULT_USER_IMAGE,
+    email: '',
+    nickName: 'test',
+    snsAccount: '',
+    createdAt: '',
     menuCount: 0
   }
 })
